more refactoring around key/value pairs
