+++ src/IdentityModel/Client/RequestUrl.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
+using IdentityModel.Internal;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
@@ -31,7 +32,7 @@ namespace IdentityModel.Client
         /// <returns></returns>
         public string Create(object values)
         {
-            var dictionary = ObjectToDictionary(values);
+            var dictionary = ValuesHelper.ObjectToDictionary(values);
             if (dictionary == null || !dictionary.Any())
             {
                 return _baseUrl;
@@ -39,29 +40,6 @@ namespace IdentityModel.Client
 
             var qs = string.Join("&", dictionary.Select(kvp => string.Format("{0}={1}", WebUtility.UrlEncode(kvp.Key), WebUtility.UrlEncode(kvp.Value))).ToArray());
             return string.Format("{0}?{1}", _baseUrl, qs);
-        }
-
-        private static Dictionary<string, string> ObjectToDictionary(object values)
-        {
-            if (values == null)
-            {
-                return null;
-            }
-
-            if (values is Dictionary<string, string> dictionary) return dictionary;
-
-            dictionary = new Dictionary<string, string>();
-
-            foreach (var prop in values.GetType().GetRuntimeProperties())
-            {
-                var value = prop.GetValue(values) as string;
-                if (!string.IsNullOrEmpty(value))
-                {
-                    dictionary.Add(prop.Name, value);
-                }
-            }
-
-            return dictionary;
         }
     }
 }
