+++ src/IdentityModel/Client/RequestUrlExtensions.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
+using IdentityModel.Internal;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
@@ -20,7 +21,7 @@ namespace IdentityModel.Client
         /// <returns></returns>
         public static string Create(this RequestUrl request, object values)
         {
-            return request.Create(ObjectToDictionary(values));
+            return request.Create(ValuesHelper.ObjectToDictionary(values));
         }
 
         /// <summary>
@@ -62,94 +63,18 @@ namespace IdentityModel.Client
                 { OidcConstants.AuthorizeRequest.ResponseType, responseType }
             };
 
-            if (!string.IsNullOrWhiteSpace(scope))
-            {
-                values.Add(OidcConstants.AuthorizeRequest.Scope, scope);
-            }
+            values.AddIfPresent(OidcConstants.AuthorizeRequest.Scope, scope);
+            values.AddIfPresent(OidcConstants.AuthorizeRequest.RedirectUri, redirectUri);
+            values.AddIfPresent(OidcConstants.AuthorizeRequest.State, state);
+            values.AddIfPresent(OidcConstants.AuthorizeRequest.Nonce, nonce);
+            values.AddIfPresent(OidcConstants.AuthorizeRequest.LoginHint, loginHint);
+            values.AddIfPresent(OidcConstants.AuthorizeRequest.AcrValues, acrValues);
+            values.AddIfPresent(OidcConstants.AuthorizeRequest.Prompt, prompt);
+            values.AddIfPresent(OidcConstants.AuthorizeRequest.ResponseMode, responseMode);
+            values.AddIfPresent(OidcConstants.AuthorizeRequest.CodeChallenge, codeChallenge);
+            values.AddIfPresent(OidcConstants.AuthorizeRequest.CodeChallengeMethod, codeChallengeMethod);
 
-            if (!string.IsNullOrWhiteSpace(redirectUri))
-            {
-                values.Add(OidcConstants.AuthorizeRequest.RedirectUri, redirectUri);
-            }
-
-            if (!string.IsNullOrWhiteSpace(state))
-            {
-                values.Add(OidcConstants.AuthorizeRequest.State, state);
-            }
-
-            if (!string.IsNullOrWhiteSpace(nonce))
-            {
-                values.Add(OidcConstants.AuthorizeRequest.Nonce, nonce);
-            }
-
-            if (!string.IsNullOrWhiteSpace(loginHint))
-            {
-                values.Add(OidcConstants.AuthorizeRequest.LoginHint, loginHint);
-            }
-
-            if (!string.IsNullOrWhiteSpace(acrValues))
-            {
-                values.Add(OidcConstants.AuthorizeRequest.AcrValues, acrValues);
-            }
-
-            if (!string.IsNullOrWhiteSpace(prompt))
-            {
-                values.Add(OidcConstants.AuthorizeRequest.Prompt, prompt);
-            }
-
-            if (!string.IsNullOrWhiteSpace(responseMode))
-            {
-                values.Add(OidcConstants.AuthorizeRequest.ResponseMode, responseMode);
-            }
-
-            if (!string.IsNullOrWhiteSpace(codeChallenge))
-            {
-                values.Add(OidcConstants.AuthorizeRequest.CodeChallenge, codeChallenge);
-            }
-
-            if (!string.IsNullOrWhiteSpace(codeChallengeMethod))
-            {
-                values.Add(OidcConstants.AuthorizeRequest.CodeChallengeMethod, codeChallengeMethod);
-            }
-
-            return request.Create(Merge(values, ObjectToDictionary(extra)));
-        }
-
-        private static Dictionary<string, string> ObjectToDictionary(object values)
-        {
-            if (values == null)
-            {
-                return null;
-            }
-
-            if (values is Dictionary<string, string> dictionary) return dictionary;
-
-            dictionary = new Dictionary<string, string>();
-
-            foreach (var prop in values.GetType().GetRuntimeProperties())
-            {
-                var value = prop.GetValue(values) as string;
-                if (!string.IsNullOrEmpty(value))
-                {
-                    dictionary.Add(prop.Name, value);
-                }
-            }
-
-            return dictionary;
-        }
-
-        private static Dictionary<string, string> Merge(Dictionary<string, string> explicitValues, Dictionary<string, string> additionalValues = null)
-        {
-            var merged = explicitValues;
-
-            if (additionalValues != null)
-            {
-                merged =
-                    explicitValues.Concat(additionalValues.Where(add => !explicitValues.ContainsKey(add.Key)))
-                                         .ToDictionary(final => final.Key, final => final.Value);
-            }
-
-            return merged;
+            return request.Create(ValuesHelper.Merge(values, ValuesHelper.ObjectToDictionary(extra)));
         }
     }
 }
